+++ src/Commands/Make/HelperMakeCommand.php
@@ -1,0 +1,77 @@
+<?php
+
+namespace Nwidart\Modules\Commands\Make;
+
+use Illuminate\Support\Str;
+use Nwidart\Modules\Support\Config\GenerateConfigReader;
+use Nwidart\Modules\Support\Stub;
+use Nwidart\Modules\Traits\ModuleCommandTrait;
+use Symfony\Component\Console\Input\InputArgument;
+use Symfony\Component\Console\Input\InputOption;
+
+class HelperMakeCommand extends GeneratorCommand
+{
+    use ModuleCommandTrait;
+
+    protected $argumentName = 'name';
+    protected $name = 'module:make-helper';
+    protected $description = 'Create a new helper class for the specified module.';
+
+    public function getDestinationFilePath(): string
+    {
+        $path = $this->laravel['modules']->getModulePath($this->getModuleName());
+
+        $filePath = GenerateConfigReader::read('helpers')->getPath() ?? config('modules.paths.app_folder') . 'Helpers';
+
+        return $path . $filePath . '/' . $this->getHelperName() . '.php';
+    }
+
+    protected function getTemplateContents(): string
+    {
+        $module = $this->laravel['modules']->findOrFail($this->getModuleName());
+
+        return (new Stub($this->getStubName(), [
+            'CLASS_NAMESPACE'   => $this->getClassNamespace($module),
+            'CLASS'             => $this->getClassNameWithoutNamespace(),
+        ]))->render();
+    }
+
+    protected function getArguments(): array
+    {
+        return [
+            ['name', InputArgument::REQUIRED, 'The name of the helper class.'],
+            ['module', InputArgument::OPTIONAL, 'The name of module will be used.'],
+        ];
+    }
+
+    /**
+     * @return array
+     */
+    protected function getOptions(): array
+    {
+        return [
+            ['invokable', 'i', InputOption::VALUE_NONE, 'Generate an invokable class', null],
+            ['force', 'f', InputOption::VALUE_NONE, 'su.'],
+        ];
+    }
+
+    protected function getHelperName(): array|string
+    {
+        return Str::studly($this->argument('name'));
+    }
+
+    private function getClassNameWithoutNamespace(): array|string
+    {
+        return class_basename($this->getHelperName());
+    }
+
+    public function getDefaultNamespace(): string
+    {
+        return config('modules.paths.generator.helpers.namespace', 'Helpers');
+    }
+
+    protected function getStubName(): string
+    {
+        return $this->option('invokable') === true ? '/helper-invoke.stub' : '/helper.stub';
+    }
+}