+++ src/Contracts/ConfirmableCommand.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Nwidart\Modules\Contracts;
+
+interface ConfirmableCommand
+{
+}